feat: listing without filters and pagination wip
