--- src/components/Table/TableFooterMobile.tsx
@@ -1,72 +1,0 @@
-import Image from "next/image";
-import { useRouter } from "next/router";
-import ArrowIcon from "public/open_input_black.svg";
-import { ICON_SIZES } from "~/utils";
-import type { ITableFooter } from "./table.interface";
-import { nextPage, previousPage, selectAnyPage } from "./utils";
-
-export const TableFooterMobile = ({
-  pagination,
-  setPagination,
-}: ITableFooter) => {
-  const router = useRouter();
-  const { page } = pagination;
-
-  const handlePrevious = () => previousPage(setPagination, pagination, router);
-  const handleNext = () => nextPage(setPagination, pagination, router);
-  const handleSelectPage = (page: number) => () =>
-    selectAnyPage(setPagination, page, router);
-
-  return (
-    <div className="flex w-full items-center justify-center gap-4 rounded-lg bg-white py-4 md:hidden">
-      <button
-        onClick={handlePrevious}
-        disabled={page === 1}
-        className="flex h-8 cursor-pointer select-none items-center justify-center rounded-lg p-2 hover:ring-1 hover:ring-m-purple/10 disabled:cursor-default disabled:opacity-50 disabled:ring-0"
-      >
-        <Image
-          src={ArrowIcon as string}
-          alt="<"
-          className="rotate-90"
-          height={ICON_SIZES.S}
-        />
-        <p>Previous</p>
-      </button>
-
-      <div className="flex gap-2">
-        {page && page >= 2 && (
-          <div
-            className="flex h-8 w-8 cursor-pointer items-center justify-center rounded-lg p-2 hover:ring-1 hover:ring-m-purple/10"
-            onClick={handleSelectPage(page - 1)}
-          >
-            <p>{page - 1}</p>
-          </div>
-        )}
-        <div className="flex h-8 w-8 cursor-pointer items-center justify-center rounded-lg bg-m-purple p-2 text-white">
-          <p>{page}</p>
-        </div>
-        {pagination.fetchMore && page && (
-          <div
-            className="flex h-8 w-8 cursor-pointer items-center justify-center rounded-lg p-2 hover:ring-1 hover:ring-m-purple/10"
-            onClick={handleSelectPage(page + 1)}
-          >
-            <p>{page + 1}</p>
-          </div>
-        )}
-      </div>
-
-      <button
-        onClick={handleNext}
-        className="flex h-8 cursor-pointer select-none items-center justify-center rounded-lg p-2 hover:ring-1 hover:ring-m-purple/10 disabled:cursor-default disabled:opacity-50 disabled:ring-0"
-      >
-        <p>Next</p>
-        <Image
-          src={ArrowIcon as string}
-          alt=">"
-          className="-rotate-90"
-          height={ICON_SIZES.S}
-        />
-      </button>
-    </div>
-  );
-};
